refactor: Move tests to separate unittest file So that IDEs can pick it up properly etc
